Progress on MeshCombine
We still need to delete children and fix a collider that matches the
parent.

Later, we will start creating containers for about 1000 cubes, each
will have a parent, they will be created within for-loops so that this
algorithm works for more pixels than the few we have now.

Should work, hope so! :)

Best regards,
Emil

IT specialist, maker of love, destroyer of gluten